+++ src/Commands/Model/AzureADApp.cs
@@ -7,12 +7,16 @@ using System.Text.Json.Serialization;
 
 namespace PnP.PowerShell.Commands.Model
 {
-    internal class AzureApp
+    public class AzureADApp
     {
         public string Id { get; set; }
         public string AppId { get; set; }
         public string DisplayName { get; set; }
         public string SignInAudience { get; set; }
+        public bool? IsFallbackPublicClient { get; set; }
+        public List<AppResource> RequiredResourceAccess { get; set; }
+
+        public AppPublicClient PublicClient { get; set; }
     }
 
     public class AppResource
@@ -21,5 +25,10 @@ namespace PnP.PowerShell.Commands.Model
         public string Id { get; set; }
         [JsonPropertyName("resourceAccess")]
         public List<PermissionScope> ResourceAccess { get; set; } = new List<PermissionScope>();
+    }
+
+    public class AppPublicClient
+    {
+        public string[] RedirectUris { get; set; }
     }
 }