+++ src/Commands/Model/PermissionScope.cs
@@ -6,10 +6,13 @@ namespace PnP.PowerShell.Commands.Model
     {
         [JsonIgnore]
         public string resourceAppId { get; set; }
+
         [JsonPropertyName("id")]
         public string Id { get; set; }
+
         [JsonPropertyName("type")]
         public string Type { get; set; } = "Role";
+        
         [JsonIgnore]
         public string Identifier { get; set; }
     }