+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Current Nightly]
 
 ### Added
+- Added Get-PnPAzureADApp, Get-PnPAzureADAppPermission and Remove-PnPAzureADApp to manage Azure AD apps.
 - Added All Graph permissions and all SharePoint permissions for selection to Register-PnPAzureADApp.
 - Added `-Template` parameter to New-PnPTeamsTeam to create teams with EDU templates (your tenant needs an EDU license)
 - Added fixes for authentication to GCC, GCC High and GCC DoD environments using certificate or interactive login.